+++ common/src/main/java/com/github/smallinger/copperagebackport/CommonClass.java
@@ -20,6 +20,7 @@ public class CommonClass {
         // Register all content
         ModSounds.register();
         ModMemoryTypes.register();
+        ModParticles.register();
         ModBlocks.register();
         ModBlockEntities.register();
         ModEntities.register();